requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
python-dateutil>=2.8.2
matplotlib>=3.7.0
pandas>=2.0.0
streamlit>=1.20.0 